converted js to ts, ommitted better ts types from rust
