+++ .gitignore
@@ -22,3 +22,9 @@
 # Node
 /node_modules
 *.log
+
+# TypeScript compilation output
+index.js
+index.d.ts
+index.js.map
+index.d.ts.map
