+++ Cargo.toml
@@ -13,7 +13,5 @@ crate-type = ["cdylib"]
 wasm-bindgen = "0.2"
 serde = { version = "1.0", features = ["derive"] }
 serde-wasm-bindgen = "0.6"
-
-[dependencies.web-sys]
-version = "0.3"
-features = ["console"]
+tsify = { version = "0.5", features = ["js"] }
+web-sys = { version = "0.3", features = ["console"] }
