+++ tsconfig.json
@@ -1,0 +1,33 @@
+{
+	"compilerOptions": {
+		"target": "ES2022",
+		"module": "commonjs",
+		"lib": ["ES2022"],
+		"outDir": "./",
+		"rootDir": "./",
+		"strict": true,
+		"esModuleInterop": true,
+		"skipLibCheck": true,
+		"forceConsistentCasingInFileNames": true,
+		"resolveJsonModule": true,
+		"declaration": true,
+		"declarationMap": true,
+		"sourceMap": true,
+		"moduleResolution": "node",
+		"allowJs": true,
+		"checkJs": false,
+		"noEmitOnError": true,
+		"removeComments": false,
+		"preserveConstEnums": true
+	},
+	"include": ["index.ts"],
+	"exclude": [
+		"node_modules",
+		"pkg",
+		"tests",
+		"**/*.test.ts",
+		"**/*.spec.ts",
+		"index.js",
+		"index.d.ts"
+	]
+}
